requirements.txt
# Core ML libraries
torch >=2.2.0
torchvision >=0.15.1
scikit-learn ==1.6.1
pandas >=2.1.0
numpy >=2.1.0
imbalanced-learn >=0.11.0

# Visualization (needed for enhanced reports)
matplotlib >=3.5.0
seaborn >=0.11.2

# Explainability
shap >=0.41.0
captum >=0.6.1

# API framework
flask >=2.3.0
flask-cors >=3.0.10

# Monitoring
prometheus-client >=0.21.0
structlog >=24.4.0

# Data processing
joblib >=1.4.2

# Testing
pytest >=7.0.0
pytest-cov >=3.0.0

# Deployment
gunicorn >=20.1.0

# Utilities
python-dotenv >=1.0.0
pyyaml >=5.4.1
psutil >=5.8.0

# Data validation and monitoring
evidently >=0.2.0

# Additional dependencies for enhanced ML pipeline
pyarrow >=8.0.0    # Latest for parquet files
tqdm >=4.64.0      # Progress bars
loguru >=0.5.3
pathlib >=1.0.1

# Security libraries
cryptography >=43.0.3
urllib3 >=2.2.3

# Performance optimization
numba >=0.60.0          # JIT compilation
bottleneck >=1.4.2      # Fast NumPy array functions

prometheus-client >=0.21.0
structlog >=24.4.0

typing-extensions >=4.10.0